requirements.txt
faiss-cpu 
pymupdf 
pdfplumber
pytesseract 
pillow 
streamlit 
numpy 
tqdm 
requests